fix bugs

+++ vector_utils.py
@@ -6,13 +6,10 @@ import openai
 import chromadb
 from chromadb.config import Settings
 from chromadb.utils import embedding_functions
-from dotenv import load_dotenv
 import os
-from llm_utils import get_embedding
+from config import OPEN_AI_API_KEY, DEV_MODE
 
 # SET UP OPENAI AND THE VECTORDB CLIENT AND COLLECTION
-load_dotenv()
-OPEN_AI_API_KEY = os.getenv("OPEN_AI_API_KEY")
 openai.api_key = OPEN_AI_API_KEY
 
 openai_ef = embedding_functions.OpenAIEmbeddingFunction(
@@ -26,9 +23,11 @@ client = chromadb.Client(Settings(
 ))
 
 try:
-  #collection = client.get_collection(name='memories')
-  client.delete_collection(name="memories")
-  collection = client.create_collection(name='memories', embedding_function=openai_ef)
+  if DEV_MODE:
+    client.delete_collection(name="memories")
+    collection = client.create_collection(name='memories', embedding_function=openai_ef)
+  else:
+    collection = client.get_collection(name='memories', embedding_function=openai_ef)
 except ValueError as e:
   collection = client.create_collection(name='memories', embedding_function=openai_ef)
 
@@ -38,15 +37,14 @@ def store_memory_in_vectordb(agent_name: str, memory: Memory) -> None:
   '''
   collection.add(
       documents=[memory.description],
-      #embeddings = [memory.embedding],
       metadatas=[{"agent": agent_name, "created_at": str(memory.created_at), "last_accessed": str(memory.last_accessed), "importance_score": memory.importance_score}],
       ids=[str(uuid.uuid4())]
   )
 
 def similiarty_search(agent_name: str, query: str, n_results: int = 5) -> List:
-  #query_embeddings = get_embedding(query)
-  collection.query(
+  results = collection.query(
     query_texts=[query],
     n_results=n_results,
     where={"agent": agent_name}
   )
+  return results
